LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-location-expected.txt


FAIL location API fires navigate event assert_equals: expected "replace" but got "push"
